Split massive function into many different descriptive tasks.
