+++ settings.gradle
@@ -1,5 +1,1 @@
 rootProject.name = 'pogo.scraper'
-rootProject.name = 'pogo.scraper'
-
-/*include ':PokeGOAPI-Java'
-project(':PokeGOAPI-Java').projectDir = file('../PokeGOAPI-Java')*/
