+++ src/main/kotlin/ink/abb/pogo/scraper/Context.kt
@@ -3,9 +3,10 @@ package ink.abb.pogo.scraper
 import com.google.common.util.concurrent.AtomicDouble
 import com.pokegoapi.api.PokemonGo
 import com.pokegoapi.api.player.PlayerProfile
+import java.util.concurrent.atomic.AtomicBoolean
 
 /**
- * @author Andrew Potter (ddcapotter)
+ * @author Andrew Potter (apottere)
  */
 
 data class Context(
@@ -14,5 +15,5 @@ data class Context(
     val lat: AtomicDouble,
     val lng: AtomicDouble,
 
-    var walking: Boolean = false
+    var walking: AtomicBoolean = AtomicBoolean(false)
 )
